refactor: split data access code from resolvers and routes
