+++ config/default.js
@@ -10,6 +10,7 @@ module.exports = {
   pubsweet: {
     components: [
       '@elifesciences/xpub-meca-export',
+      '@elifesciences/xpub-model',
       '@elifesciences/xpub-server',
       '@elifesciences/xpub-client-config',
       '@pubsweet/component-send-email',
