+++ server/meca/file-generators/article.js
@@ -1,7 +1,7 @@
 const xmlbuilder = require('xmlbuilder')
 const lodash = require('lodash')
 const config = require('config')
-const elifeApi = require('@elifesciences/xpub-server/entities/user/helpers/elife-api')
+const { UserManager } = require('@elifesciences/xpub-model')
 const {
   articleTypeMap,
   contribStructure,
@@ -110,7 +110,7 @@ async function generateArticleXml(manuscript) {
     return ids
   }, [])
 
-  const editors = await elifeApi.peopleById(editorIds)
+  const editors = await UserManager.getEditorsByPersonId(editorIds)
   const editorsById = editors.reduce(
     (accumulator, editor) => ({ ...accumulator, [editor.id]: editor }),
     {},
