+++ server/meca/package.json
@@ -8,6 +8,7 @@
   "author": "eLife",
   "license": "MIT",
   "dependencies": {
+    "@elifesciences/xpub-model": "1.0.0",
     "aws-sdk": "^2.313.0",
     "fs-extra": "^7.0.0",
     "jszip": "^3.1.5",
