+++ server/xpub-model/entities/manuscript/helpers/empty.js
