+++ server/xpub-model/entities/manuscript/index.js
