+++ server/xpub-model/index.js
@@ -1,0 +1,13 @@
+const FileManager = require('./entities/file')
+const IdentityManager = require('./entities/identity')
+const ManuscriptManager = require('./entities/manuscript')
+const TeamManager = require('./entities/team')
+const UserManager = require('./entities/user')
+
+module.exports = {
+  FileManager,
+  IdentityManager,
+  ManuscriptManager,
+  TeamManager,
+  UserManager,
+}
