+++ server/xpub-model/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@elifesciences/xpub-model",
+  "version": "1.0.0",
+  "description": "",
+  "license": "MIT",
+  "author": "",
+  "main": "index.js",
+  "scripts": {},
+  "keywords": [],
+  "dependencies": {
+    "config": "^2.0.1",
+    "fs-extra": "^7.0.0",
+    "lodash": "^4.17.5",
+    "request": "^2.88.0",
+    "request-promise-native": "^1.0.5",
+    "superagent": "^3.8.2",
+    "uuid": "^3.3.2",
+    "xml2js": "^0.4.19"
+  },
+  "peerDependencies": {
+    "pubsweet-server": ">=2.0.0"
+  }
+}
