+++ server/xpub-server/entities/manuscript/typeDefs.graphqls
